Added scripts to new folder
