+++ Master Project/Assets/Scenes/Shaking/Assets.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: d0d2fb49600bac74683b42c3be7f7688
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
