+++ Master Project/Assets/Scenes/Shaking/Scripts.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 82bc60842f24e4d4e86ada1d489d4592
+guid: 5aeda4208d3a25b409f9730b9680c6d8
 folderAsset: yes
 DefaultImporter:
   externalObjects: {}
